cleanup products cards
